backend/requirements.txt
Django==4.1.6
djangorestframework==3.13.1
drf-spectacular==0.23.1
django-filter==22.1
python-decouple==3.7
psycopg2==2.9.5